pyproject.toml
[tool.poetry]
name = "shop-bot"
version = "0.1.0"
description = "Telegram shop with admin panel."
authors = ["Ilya Kuchmenko <42113015+ilyahahaha@users.noreply.github.com>"]
packages = [{include = "shop_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = {version = "^3.0.0rc1", allow-prereleases = true}
aiohttp = "^3.8.5"
aiohttp-jinja2 = "^1.5.1"
aiohttp-session = {extras = ["secure"], version = "^2.12.0"}
pydantic = "^2.1.1"
pydantic-settings = "^2.0.2"
sqlalchemy = "^2.0.19"
asyncpg = "^0.28.0"
alembic = "^1.11.2"
cuid2 = "^2.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pillow = "^10.0.0"
aiofiles = "^23.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.284"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
